deployment/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
transformers==4.40.1
huggingface_hub==0.23.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
torch==2.2.2
gradio==4.44.1
pydantic>=2.6
numpy
sentencepiece>=0.1.99
protobuf>=3.20.3
bitsandbytes==0.43.1
accelerate==0.25.0
scipy==1.11.4                               